+++ htmldoc/ps-pdf.cxx
@@ -457,11 +457,14 @@ pspdf_export(tree_t *document,	/* I - Document to export */
     logo_width  = (float)(logo_image->width * PagePrintWidth / _htmlBrowserWidth);
     logo_height = (float)(logo_width * logo_image->height / logo_image->width);
 
-    if (logo_height > HeadFootSize)
+    if (logo_height > (2.0 * HeadFootSize))
     {
       // Issue #273: too large logo image will overlap the body text, so cap
       // the height of the logo image to the header/footer size...
-      logo_height = (float)HeadFootSize;
+      //
+      // Issue #303: regression prevents using header/footer images for special
+      // underlining/etc. effects.
+      logo_height = (float)(2.0 * HeadFootSize);
       logo_width  = logo_height * logo_image->width / logo_image->height;
     }
 
@@ -480,11 +483,14 @@ pspdf_export(tree_t *document,	/* I - Document to export */
       hfimage_width[hfi]  = (float)(hfimage[hfi]->width * PagePrintWidth / _htmlBrowserWidth);
       hfimage_height[hfi] = (float)(hfimage_width[hfi] * hfimage[hfi]->height / hfimage[hfi]->width);
 
-      if (hfimage_height[hfi] > HeadFootSize)
+      if (hfimage_height[hfi] > (2.0 * HeadFootSize))
       {
         // Issue #273: too large logo image will overlap the body text, so cap
         // the height of the logo image to the header/footer size...
-        hfimage_height[hfi] = (float)HeadFootSize;
+        //
+        // Issue #303: regression prevents using header/footer images for special
+        // underlining/etc. effects.
+        hfimage_height[hfi] = (float)(2.0 * HeadFootSize);
         hfimage_width[hfi]  = hfimage_height[hfi] * hfimage[hfi]->width / hfimage[hfi]->height;
       }